# openapi.yaml
openapi: 3.0.3
info:
  title: Virly API
  version: 1.0.0
  description: |
    Swagger-compatible OpenAPI document for the Virly backend.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Creates a pending user account, generates a short-lived verification
        link, and sends it by email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              default:
                value:
                  email: user@example.com
                  password: StrongPass123!
                  rememberMe: true
                  phone: "+15551234567"
      responses:
        '201':
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/verify:
    get:
      tags:
        - Authentication
      summary: Verify email using a verification link token
      description: Confirms the user's email, activates the account, and sets auth cookies.
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
          description: Short-lived email verification token from the verification link
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Invalid or expired verification token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend an email verification link
      description: |
        Generates a fresh short-lived verification link for a registered,
        unverified user and sends it by email. Unknown email addresses and
        already verified accounts receive the same generic success response to
        avoid revealing account existence or verification state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
            examples:
              default:
                value:
                  email: user@example.com
      responses:
        '200':
          description: Verification email sent or request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Log in with email and password
      description: Sets the HttpOnly auth cookie and readable CSRF cookie on success, and returns the CSRF token for cross-origin frontends that cannot read the API-domain cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              default:
                value:
                  email: user@example.com
                  password: StrongPass123!
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account is not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Log out the current user
      description: Clears the auth and CSRF cookies.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfTokenHeader'
      responses:
        '200':
          description: Logout acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Missing or invalid auth cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get the current authenticated user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          description: Missing or invalid auth cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
...
    CsrfTokenHeader:
      in: header
      name: X-CSRF-Token
      required: true
      schema:
        type: string
      description: Value from the csrfToken auth response field or readable virly_csrf cookie. Required for authenticated unsafe methods.
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: virly_auth
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - phone
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: StrongPass123!
        phone:
          type: string
          description: E.164 formatted phone number
          example: "+15551234567"
    RegisterResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Verification email sent to user@example.com
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        rememberMe:
          type: boolean
          default: false
          description: When true, auth and CSRF cookies are persistent for 30 days and refreshed on each successful login. When false or omitted, browser-session cookies are used.
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    AuthSuccessResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
        csrfToken:
          type: string
          description: CSRF token bound to the auth cookie. Same-origin clients can also read this from the virly_csrf cookie; cross-origin clients should cache this response value and send it in X-CSRF-Token for unsafe methods.
          example: lRZlG2Yd3xEY3EbOg8p7Yj37TyoEL7x9ySd9uSx4vxM
    User:
      type: object
      required:
        - id
        - email
        - balance
        - personalDetailsId
        - personalDetailsStatus
        - needsPersonalDetails
      properties:
        id:
          type: string
          example: 6630c4b2e93e5f6c8a321001
        email:
          type: string
          format: email
          example: user@example.com
        balance:
          type: number
          format: double
          example: 1042.75
        personalDetailsId:
          type: string
          example: 6630c4b2e93e5f6c8a321010
        personalDetailsStatus:
          type: string
          enum:
            - not_provided
            - provided
          example: not_provided
        needsPersonalDetails:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-04-28T09:15:00.000Z"
...
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Logged out successfully
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Invalid credentials
        details:
          type: array
          items:
            type: string
        issues:
          type: array
          items:
            type: object
            required:
              - path
              - message
            properties:
              path:
                type: string
                example: email
              message:
                type: string
                example: Invalid email